requirements.txt
pyserial
python-socketio
requests
OPi.GPIO
